requirements.txt
Flask>=3.0.0
flask-limiter>=3.5.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
Werkzeug>=3.0.0
gunicorn>=21.2.0
